Extracting tlsloader into a separate package to enable reuse
